requirements.txt
absl-py==1.4.0
accelerate==0.21.0
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.0.1
anyio==3.6.2
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
bitsandbytes==0.40.2
blinker==1.6.2
blis==0.7.10
branca==0.6.0
cachetools==5.3.1
catalogue==2.0.9
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
confection==0.1.1
contourpy==1.1.0
ctransformers==0.2.27
cycler==0.11.0
cymem==2.0.7
dataclasses-json==0.6.1
datasets==2.14.4
dill==0.3.7
docopt==0.6.2
exceptiongroup==1.1.2
faiss-cpu==1.7.4
fake-useragent==1.1.3
fastapi==0.95.2
ffmpy==0.3.1
filelock==3.12.2
Flask==2.3.2
flatbuffers==23.5.26
folium==0.14.0
fonttools==4.42.0
frozenlist==1.4.0
fsspec==2023.6.0
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.37
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
gradio==3.40.1
gradio_client==0.4.0
grpcio==1.57.0
h11==0.14.0
h5py==3.9.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.16.4
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
jsonify==0.5
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
keras==2.13.1
kiwisolver==1.4.4
langchain==0.0.309
langcodes==3.3.0
langsmith==0.0.42
libclang==16.0.6
linkify-it-py==2.0.2
Markdown==3.4.4
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.20.1
matplotlib==3.7.2
mdit-py-plugins==0.3.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
murmurhash==1.0.9
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numpy==1.24.3
oauthlib==3.2.2
openai==0.27.9
opencv-python==4.8.0.76
opt-einsum==3.3.0
orjson==3.9.5
outcome==1.2.0
packaging==23.1
pandas==2.0.3
pathy==0.10.2
peft==0.4.0
Pillow==10.0.0
preshed==3.0.8
protobuf==4.24.1
psutil==5.9.5
py-cpuinfo==9.0.0
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.8
pydeck==0.8.1b0
pydub==0.25.1
Pygments==2.16.1
pyparsing==3.0.9
pypdf==3.16.2
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.8
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rich==13.6.0
rpds-py==0.9.2
rsa==4.9
safetensors==0.3.2
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
selenium==4.10.0
semantic-version==2.10.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smart-open==6.3.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==2.0.21
srsly==2.4.7
starlette==0.27.0
streamlit==1.27.1
streamlit-chat==0.1.1
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-macos==2.13.0
termcolor==2.3.0
thinc==8.1.12
threadpoolctl==3.1.0
tiktoken==0.5.1
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torch-tb-profiler==0.4.1
torchaudio==2.0.2
torchvision==0.15.2
tornado==6.3.3
tqdm==4.65.0
transformers==4.31.0
trio==0.22.1
trio-websocket==0.10.3
trl==0.4.7
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==5.0.1
uc-micro-py==1.0.2
ultralytics==8.0.188
urllib3==2.0.2
uvicorn==0.22.0
validators==0.22.0
wasabi==1.1.2
websockets==11.0.3
Werkzeug==2.3.4
wrapt==1.15.0
wsproto==1.2.0
xxhash==3.3.0
yarl==1.9.2
zipp==3.17.0
